add Notebooks component to contain importing demo notebooks
