+++ src/taskpane/components/FunctionsTab.jsx
@@ -1,24 +1,12 @@
 import * as React from "react";
 import { getFunctionFromSettings, deleteFunctionFromSettings } from "../utils/workbookSettings";
-import { DEFAULT_NOTEBOOK, getStoredNotebooks, addNotebook, removeNotebook, fetchDemoNotebooks, fetchNotebookUrl } from "../utils/notebooks";
-import { saveFunctionToSettings } from "../utils/workbookSettings";
-import { updateNameManager } from "../utils/nameManager";
 import { runTests } from "../utils/testRunner";
+import Notebooks from "./Notebooks";
 
 const FunctionsTab = ({ onEdit, onTest }) => {
-    // Original state for functions
     const [functions, setFunctions] = React.useState([]);
     const [error, setError] = React.useState(null);
-
-    // New state for notebooks
-    const [myNotebooks, setMyNotebooks] = React.useState({});
-    const [url, setUrl] = React.useState('');
-    const [isUrlSaving, setIsUrlSaving] = React.useState(false);
-    const [saveSuccess, setSaveSuccess] = React.useState(false);
-    const [selectedNotebook, setSelectedNotebook] = React.useState(DEFAULT_NOTEBOOK);
-    const [isImporting, setIsImporting] = React.useState(false);
     const [deleteConfirm, setDeleteConfirm] = React.useState(null);
-    const [demoNotebooks, setDemoNotebooks] = React.useState({});
 
     const loadFunctions = async () => {
         try {
@@ -30,79 +18,9 @@ const FunctionsTab = ({ onEdit, onTest }) => {
         }
     };
 
-    const loadNotebooks = async () => {
-        try {
-            const storedNotebooks = await getStoredNotebooks();
-            setMyNotebooks(storedNotebooks);
-        } catch (error) {
-            console.error('Error loading notebooks:', error);
-            setError('Failed to load notebooks. Please try again.');
-        }
-    };
-
-    // Combined loading effect
     React.useEffect(() => {
-        const loadData = async () => {
-            try {
-                await Promise.all([
-                    loadFunctions(),
-                    loadNotebooks(),
-                    fetchDemoNotebooks().then(setDemoNotebooks)
-                ]);
-            } catch (error) {
-                console.error('Error loading data:', error);
-                setError('Failed to load data. Please try again.');
-            }
-        };
-
-        loadData();
+        loadFunctions();
     }, []);
-
-    const handleUrlSubmit = async (e) => {
-        e.preventDefault();
-        setIsUrlSaving(true);
-        setSaveSuccess(false);
-        try {
-            await addNotebook(url);
-            await loadNotebooks();
-            setUrl('');
-            setSaveSuccess(true);
-            setTimeout(() => setSaveSuccess(false), 3000);
-        } catch (error) {
-            setError(`Failed to add notebook: ${error.message}`);
-        } finally {
-            setIsUrlSaving(false);
-        }
-    };
-
-    const handleRemove = async (url) => {
-        try {
-            await removeNotebook(url);
-            await loadNotebooks();
-        } catch (error) {
-            setError('Failed to remove notebook');
-        }
-    };
-
-    const handleImportFunctions = async () => {
-        setIsImporting(true);
-        try {
-            const { functions: parsedFunctions } = await fetchNotebookUrl(selectedNotebook);
-
-            for (const func of parsedFunctions) {
-                await saveFunctionToSettings(func);
-                await updateNameManager(func);
-            }
-
-            await loadFunctions();
-            setError(null);
-        } catch (error) {
-            console.error("Error importing functions:", error);
-            setError("Failed to import functions");
-        } finally {
-            setIsImporting(false);
-        }
-    };
 
     const handleDelete = async (functionName) => {
         try {
@@ -127,7 +45,7 @@ const FunctionsTab = ({ onEdit, onTest }) => {
 
     return (
         <div className="h-full flex flex-col overflow-y-auto">
-            {/* Consolidated Error Display */}
+            {/* Error Display */}
             {error && (
                 <div className="p-4 text-red-600 bg-red-50 mb-4">
                     {error}
@@ -187,97 +105,7 @@ const FunctionsTab = ({ onEdit, onTest }) => {
             </div>
 
             {/* Notebooks Section */}
-            <div className="border-t pt-2">
-                <div className="px-4 mb-4">
-                    <select
-                        value={selectedNotebook}
-                        onChange={(e) => setSelectedNotebook(e.target.value)}
-                        className="w-full p-2 border border-gray-300 rounded-lg text-sm mb-2"
-                    >
-                        <option value="">Select a notebook with example functions...</option>
-                        {Object.keys(myNotebooks).length > 0 && (
-                            <optgroup label="Custom Notebooks">
-                                {Object.entries(myNotebooks).map(([url, { fileName }]) => (
-                                    <option key={url} value={url}>{fileName || url}</option>
-                                ))}
-                            </optgroup>
-                        )}
-                        {Object.entries(demoNotebooks).map(([category, notebooks]) => (
-                            <optgroup key={category} label={category}>
-                                {notebooks.map(({ url, title }) => (
-                                    <option key={url} value={url}>{title}</option>
-                                ))}
-                            </optgroup>
-                        ))}
-                    </select>
-                    <button
-                        onClick={handleImportFunctions}
-                        disabled={isImporting || !selectedNotebook}
-                        className="w-full px-3 py-2 bg-blue-500 text-white rounded-lg hover:bg-blue-600 disabled:bg-blue-300"
-                    >
-                        {isImporting ? "Importing Functions..." : "Import Notebook Functions"}
-                    </button>
-                </div>
-
-                <div className="px-4 mb-2">
-                    <form onSubmit={handleUrlSubmit} className="mb-2">
-                        <div className="flex gap-2 items-center">
-                            <input
-                                type="url"
-                                value={url}
-                                onChange={(e) => setUrl(e.target.value)}
-                                className="flex-1 px-2 py-1 border rounded"
-                                placeholder="Enter notebook URL (advanced)"
-                            />
-                            <button
-                                type="submit"
-                                disabled={isUrlSaving || !url.trim()}
-                                className="px-3 py-1 bg-gray-500 text-white rounded hover:bg-gray-700 disabled:opacity-50"
-                            >
-                                {isUrlSaving ? 'Adding...' : 'Add'}
-                            </button>
-                        </div>
-                        {saveSuccess && (
-                            <div className="mt-2 text-green-600 text-sm">
-                                Notebook added successfully!
-                            </div>
-                        )}
-                    </form>
-                </div>
-
-                {Object.keys(myNotebooks).length > 0 && (
-                    <div className="overflow-x-auto mb-2">
-                        <table className="min-w-full bg-white">
-                            <thead>
-                                <tr>
-                                    <th className="py-2 px-4 border-b text-left">Custom Notebooks</th>
-                                    <th className="py-2 px-4 border-b"></th>
-                                </tr>
-                            </thead>
-                            <tbody>
-                                {Object.entries(myNotebooks).map(([url, notebook]) => (
-                                    <tr key={url}>
-                                        <td className="py-2 px-4 border-b">
-                                            <a href={url} target="_blank" rel="noopener noreferrer"
-                                                className="text-blue-500 hover:underline">
-                                                {notebook.fileName || 'Untitled'}
-                                            </a>
-                                        </td>
-                                        <td className="py-2 px-4 border-b w-fit text-right">
-                                            <button
-                                                className="text-red-500 hover:underline"
-                                                onClick={() => handleRemove(url)}
-                                            >
-                                                Remove
-                                            </button>
-                                        </td>
-                                    </tr>
-                                ))}
-                            </tbody>
-                        </table>
-                    </div>
-                )}
-            </div>
+            <Notebooks onImportComplete={loadFunctions} />
 
             {/* Delete Confirmation Dialog */}
             {deleteConfirm && (
